move checkPhase into the run script
Currently, all-evaluators has a build-time dependency on running the
tests, but not a run-time dependency. Thus after garbage collection, a
change to one evaluator can cause every other evaluator's tests to be
run again.

This is because all-evaluators is a symlinkJoin, so at runtime all
that's left of the evaluator derivations is the "run" script, which is
untested.

We have to do chmod +x in the checkPhase for now because of a bug in
nixpkgs.